+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - (Bugfix) Enable Platform Operator on EE Chart
 - (Feature) Improve GRPC JSON Handling
 - (Bugfix) Fix Operator Pod Resources
+- (Feature) Improve Session & Route Handling
 
 ## [1.3.0](https://github.com/arangodb/kube-arangodb/tree/1.3.0) (2025-08-01)
 - (Feature) (Platform) Storage Debug